requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytz==2023.3
google-generativeai>=0.3.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9 